+++ packages/bridge-controller/src/bridge-controller.test.ts
@@ -72,6 +72,7 @@ const bridgeConfig = {
   maxRefreshCount: 3,
   refreshRate: 3,
   support: true,
+  chainRanking: [],
   chains: {
     '10': { isActiveSrc: true, isActiveDest: false },
     '534352': { isActiveSrc: true, isActiveDest: false },
@@ -2703,6 +2704,7 @@ describe('BridgeController', function () {
         enabled: true,
         minimumVersion: '13.8.0',
       },
+      chainRanking: [{ chainId: 'eip155:1' as const, name: 'Ethereum' }],
     };
 
     const quotesByDecreasingProcessingTime = [...mockBridgeQuotesSolErc20];